Improved on the equation solver design in Pyro.Math.dll, it now correctly calculates simple addition and subtraction equations.Added factorial support.
Due to the improvements in the Pyro.Math.dll mentioned above, a new feature has been added, and that is being able to insert equations into the value field of a parameter for an ICommand of *ANY* type.
An example would be doing something such as: ex. "G00 X(11.5 - 3.2) Y(-5 + 2)", which would equate to writing "G00 X8.3 Y-3"
Another example would be doing the same but without the spaces: ex. "G00 X11.5-3.2 Y-5+2", personally I prefer the clean look of the above example, but both are equally valid.

Signed-off-by: Lil Jzf <ForsakenNydoseth@hotmail.com>